refactor: move query_reverted_withdrawal_cells to rpc_client
Also rename to query_withdrawal_cells_by_block_hashes
